Added validation of the UUID in controllers

+++ src/app/controllers/category.controller.js
@@ -1,4 +1,5 @@
 const CategoryRepository = require('../repositories/category.repository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class CategoryController {
   async index(request, response) {
@@ -8,6 +9,11 @@ class CategoryController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
+
     const category = await CategoryRepository.findById(id);
 
     if (!category) {
@@ -33,6 +39,10 @@ class CategoryController {
     const { id } = request.params;
     const { name } = request.body;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
+
     const categoryExists = await CategoryRepository.findById(id);
 
     if (!categoryExists) {
@@ -50,6 +60,10 @@ class CategoryController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid category id' });
+    }
 
     await CategoryRepository.delete(id);
     return response.sendStatus(204);
